--- src/main/java/graduation/project/no18/enums/ProviderType.java
@@ -1,5 +1,0 @@
-package graduation.project.no18.enums;
-
-public enum ProviderType {
-    GOOGLE, KAKAO, NAVER, LOCAL
-}
