+++ src/main/java/graduation/project/no18/global/oauth/config/JwtConfig.java
@@ -1,0 +1,19 @@
+package graduation.project.no18.global.oauth.config;
+
+import graduation.project.no18.global.oauth.token.AuthTokenProvider;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JwtConfig {
+
+    @Value("${jwt.secret}")
+    private String secret;
+
+    @Bean
+    public AuthTokenProvider jwtProvider(){
+        return new AuthTokenProvider(secret);
+    }
+
+}
