+++ src/main/java/graduation/project/no18/global/oauth/enums/ProviderType.java
@@ -1,0 +1,5 @@
+package graduation.project.no18.global.oauth.enums;
+
+public enum ProviderType {
+    GOOGLE, KAKAO, NAVER, LOCAL
+}
