+++ src/main/java/graduation/project/no18/global/oauth/enums/RoleType.java
@@ -1,0 +1,24 @@
+package graduation.project.no18.global.oauth.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.Arrays;
+
+@Getter
+@AllArgsConstructor
+public enum RoleType {
+    MEMBER("ROLE_MEMBER", "일반 회원"),
+    ADMIN("ROLE_ADMIN", "관리자"),
+    GUEST("ROLE_GUEST", "비회원");
+
+    private final String code;
+    private final String displayName;
+
+    public static RoleType of(String code){
+        return Arrays.stream(RoleType.values())
+                .filter(r -> r.getCode().equals(code))
+                .findAny()
+                .orElse(GUEST);
+    }
+}
