+++ src/main/java/graduation/project/no18/global/oauth/filter/TokenAuthenticationFilter.java
@@ -1,0 +1,37 @@
+package graduation.project.no18.global.oauth.filter;
+
+import graduation.project.no18.global.oauth.token.AuthTokenProvider;
+import graduation.project.no18.global.oauth.utils.HeaderUtil;
+import graduation.project.no18.global.oauth.token.AuthToken;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+@Slf4j
+@RequiredArgsConstructor
+public class TokenAuthenticationFilter extends OncePerRequestFilter {
+    private final AuthTokenProvider authTokenProvider;
+    @Override
+    protected void doFilterInternal(HttpServletRequest request,
+                                    HttpServletResponse response,
+                                    FilterChain filterChain) throws ServletException, IOException {
+
+        String tokenStr = HeaderUtil.getAccessToken(request);
+        AuthToken token = authTokenProvider.converteAuthToken(tokenStr);
+
+        if(token.validate()){
+            Authentication authentication = authTokenProvider.getAuthentication(token);
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+        }
+
+        filterChain.doFilter(request, response);
+    }
+}
