+++ src/main/java/graduation/project/no18/global/oauth/type/ProviderType.java
@@ -1,0 +1,8 @@
+package graduation.project.no18.global.oauth.type;
+
+import lombok.Getter;
+
+@Getter
+public enum ProviderType {
+    GOOGLE, NAVER, KAKAO, LOCAL;
+}
