+++ src/main/java/graduation/project/no18/global/oauth/type/RoleType.java
@@ -1,0 +1,16 @@
+package graduation.project.no18.global.oauth.type;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum RoleType {
+    MEMBER("ROLE_MEMBER", "일반 유저"),
+    ADMIN("ROLE_ADMIN", "관리자");
+
+    private final String code;
+    private final String name;
+
+
+}
