--- src/main/java/graduation/project/no18/oauth/customservice/CustomOAuth2MemberService.java
@@ -1,68 +1,0 @@
-package graduation.project.no18.oauth.customservice;
-
-import graduation.project.no18.domain.member.Member;
-import graduation.project.no18.domain.member.repository.MemberRepository;
-import graduation.project.no18.enums.ProviderType;
-import graduation.project.no18.enums.RoleType;
-import graduation.project.no18.oauth.oauth2member.OAuth2MemberFactory;
-import graduation.project.no18.oauth.oauth2member.OAuth2MemberInfo;
-import graduation.project.no18.oauth.principal.MemberPrincipal;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.InternalAuthenticationServiceException;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomOAuth2MemberService extends DefaultOAuth2UserService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest memberRequest) throws OAuth2AuthenticationException {
-        OAuth2User member = super.loadUser(memberRequest);
-
-        try{
-            return this.process(memberRequest, member);
-        }catch (Exception ex){
-            ex.printStackTrace();
-            throw new InternalAuthenticationServiceException(ex.getMessage(), ex.getCause());
-        }
-    }
-
-    private OAuth2User process(OAuth2UserRequest memberRequest, OAuth2User member){
-        ProviderType providerType = ProviderType.valueOf(
-                memberRequest
-                .getClientRegistration()
-                        .getRegistrationId()
-                        .toUpperCase());
-
-        OAuth2MemberInfo memberInfo =
-                OAuth2MemberFactory.getOAuth2MemberInfo(providerType, member.getAttributes());
-
-        Member savedMember =
-                memberRepository.findMemberByEmail(memberInfo.getEmail())
-                        .orElse(createMember(memberInfo, providerType));
-
-        if(!savedMember.checkProviderType(providerType)){
-            //throw new OAuthProviderMissMatchException();
-        }
-        return MemberPrincipal.create(savedMember, member.getAttributes());
-    }
-
-    private Member createMember(OAuth2MemberInfo memberInfo, ProviderType providerType){
-        Member member = Member.builder()
-                .email(memberInfo.getEmail())
-                .nickname(memberInfo.getNickname())
-                .profileImg(memberInfo.getProfileImg())
-                .providerType(providerType)
-                .roleType(RoleType.MEMBER)
-                .build();
-
-        return memberRepository.saveAndFlush(member);
-    }
-
-}
