--- src/main/java/graduation/project/no18/oauth/handler/TokenAccessDeniedHandler.java
@@ -1,26 +1,0 @@
-package graduation.project.no18.oauth.handler;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class TokenAccessDeniedHandler implements AccessDeniedHandler {
-
-    private final HandlerExceptionResolver handlerExceptionResolver;
-    @Override
-    public void handle(HttpServletRequest request,
-                       HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        handlerExceptionResolver.resolveException(request, response, null, accessDeniedException);
-    }
-}
