--- src/main/java/graduation/project/no18/oauth/membertoken/MemberRefreshTokenRepository.java
@@ -1,12 +1,0 @@
-package graduation.project.no18.oauth.membertoken;
-
-import com.nimbusds.oauth2.sdk.token.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface MemberRefreshTokenRepository extends JpaRepository<MemberRefreshToken, Long> {
-    Optional<RefreshToken> findByMemberId(UUID memberId);
-    Optional<RefreshToken> findByMemberIdAndRefreshToken(UUID memberId, String refreshToken);
-}
