--- src/main/java/graduation/project/no18/oauth/oauth2member/NaverOAuth2MemberInfo.java
@@ -1,42 +1,0 @@
-package graduation.project.no18.oauth.oauth2member;
-
-import java.util.Map;
-
-public class NaverOAuth2MemberInfo extends OAuth2MemberInfo {
-    public NaverOAuth2MemberInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getNickname() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("nickname");
-    }
-
-    @Override
-    public String getEmail() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("email");
-    }
-
-    @Override
-    public String getProfileImg() {
-        Map<String, Object> response = (Map<String, Object>) attributes.get("response");
-
-        if (response == null) {
-            return null;
-        }
-
-        return (String) response.get("profile_img");
-    }
-}
