--- src/main/java/graduation/project/no18/oauth/oauth2member/OAuth2MemberFactory.java
@@ -1,21 +1,0 @@
-package graduation.project.no18.oauth.oauth2member;
-
-import graduation.project.no18.enums.ProviderType;
-
-import java.util.Map;
-
-public class OAuth2MemberFactory {
-
-    public static OAuth2MemberInfo getOAuth2MemberInfo(ProviderType providerType, Map<String, Object> attributes){
-        if(providerType == ProviderType.GOOGLE){
-            return new GoogleOAuth2MemberInfo(attributes);
-        }
-        if(providerType == ProviderType.NAVER){
-            return new NaverOAuth2MemberInfo(attributes);
-        }
-        if(providerType == ProviderType.KAKAO){
-            return new KakaoOAuth2MemberInfo(attributes);
-        }
-        throw new IllegalArgumentException("Invalid ProviderType");
-    }
-}
