--- src/main/java/graduation/project/no18/oauth/token/AuthToken.java
@@ -1,89 +1,0 @@
-package graduation.project.no18.oauth.token;
-
-import io.jsonwebtoken.*;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-import java.security.Key;
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import java.util.Date;
-
-@Slf4j
-@RequiredArgsConstructor
-public class AuthToken {
-
-    @Getter
-    private final String token;
-    private final Key key;
-
-    private static final String AUTHORITIES_KEY = "role";
-
-    AuthToken(String id, LocalDateTime expiry, Key key) {
-        this.key = key;
-        this.token = createAuthToken(id, expiry);
-    }
-
-    AuthToken(String id, String role, LocalDateTime expiry, Key key) {
-        this.key = key;
-        this.token = createAuthToken(id, role, expiry);
-    }
-
-    private String createAuthToken(String id, LocalDateTime expiry) {
-        return Jwts.builder()
-                .setSubject(id)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .setExpiration(Timestamp.valueOf(expiry))
-                .compact();
-    }
-
-    private String createAuthToken(String id, String role, LocalDateTime expiry) {
-        return Jwts.builder()
-                .setSubject(id)
-                .claim(AUTHORITIES_KEY, role)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .setExpiration(Timestamp.valueOf(expiry))
-                .compact();
-    }
-
-    public boolean validate(){
-        return this.getTokenClaims() != null;
-    }
-
-    public Claims getTokenClaims(){
-        try {
-            return Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody();
-        }catch (SecurityException e) {
-            log.info("Invalid JWT signature.");
-        } catch (MalformedJwtException e) {
-            log.info("Invalid JWT token.");
-        } catch (ExpiredJwtException e) {
-            log.info("Expired JWT token.");
-        } catch (UnsupportedJwtException e) {
-            log.info("Unsupported JWT token.");
-        } catch (IllegalArgumentException e) {
-            log.info("JWT token compact of handler are invalid.");
-        }
-        return null;
-    }
-
-    public Claims getExpiredTokenClaims(){
-        try{
-            Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token)
-                    .getBody();
-        }catch (ExpiredJwtException e){
-            log.info("Expired JWT token");
-            return e.getClaims();
-        }
-        return null;
-    }
-
-}
