--- src/main/java/graduation/project/no18/oauth/type/ProviderType.java
@@ -1,8 +1,0 @@
-package graduation.project.no18.oauth.type;
-
-import lombok.Getter;
-
-@Getter
-public enum ProviderType {
-    GOOGLE, NAVER, KAKAO, LOCAL;
-}
