--- src/main/java/graduation/project/no18/oauth/utils/HeaderUtil.java
@@ -1,22 +1,0 @@
-package graduation.project.no18.oauth.utils;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-public class HeaderUtil {
-    private final static String HEADER_AUTHORIZATION = "Authorization";
-    private final static String TOKEN_PREFIX = "Bearer ";
-
-    public static String getAccessToken(HttpServletRequest request) {
-        String headerValue = request.getHeader(HEADER_AUTHORIZATION);
-
-        if (headerValue == null) {
-            return null;
-        }
-
-        if (headerValue.startsWith(TOKEN_PREFIX)) {
-            return headerValue.substring(TOKEN_PREFIX.length());
-        }
-
-        return null;
-    }
-}
